Added movies and genre endpoints (#53)
- Created movies entity
- Create genre entity
- Added endpoint to retrieve movies
- Added endpoint to import movies
- Added endpoint to retrieve genres
- Added endpoint to retrieve a single genre
- Added endpoint to retrieve a single movie

+++ src/main/java/bg/guardiankiller/moviessocialapp/model/dto/TMDBMovie.java
@@ -5,7 +5,7 @@ import java.util.Set;
 
 public record TMDBMovie(
         long id,
-        Set<Integer> genres,
+        Set<Long> genres,
         Map<Language, String> overview,
         Map<Language, String> title,
         String release_date,
